+++ changelog.md
@@ -1,4 +1,4 @@
-## Resourcify 1.4.0+beta.1
+## Resourcify 1.4.0
 
 This is a beta version, there might be issues!
 Please make sure to report any bugs and/or visual anomalies on [Resourcify's GitHub](https://github.com/DeDiamondPro/Resourcify/issues) or in the [Discord](https://discord.gg/XtAuqsJWby).
@@ -6,10 +6,11 @@ Please make sure to report any bugs and/or visual anomalies on [Resourcify's Git
 ### Major new features
 - Add support for CurseForge, you can now browse and install resource packs from CurseForge. There is a new dropdown on the browse page to switch between CurseForge and modrinth.
 - Add settings GUI, allowing you to configure the default source. The config can be accessed via the mod menu (on fabric you need the `mod menu` mod to be able to access the config).
+- Support for expandable dropdowns in project descriptions have been added.
+- [MineMark](https://github.com/DeDiamondPro/MineMark) (Resourcify's markdown library) has been updated to [1.1.0](https://github.com/DeDiamondPro/MineMark/releases/1.1.0), adding support for css style tags, like text color.
 
 ### What's next for 1.4.0
 Other features that will be added in Resourcify 1.4.0 in a later beta:
-- Support expandable spoilers in project descriptions
 - Dependency management and automatic installation of dependencies
 
 ### Other changes